fix: move renderAddressToHash function to accounts, and update imports accordingly
