+++ vue-app/src/utils/accounts.ts
@@ -24,3 +24,17 @@ export async function isValidEthAddress(address: string): Promise<boolean> {
   const resolved = await mainnetProvider.resolveName(address)
   return !!resolved
 }
+
+export function renderAddressOrHash(
+  address: string,
+  digitsToShow?: number
+): string {
+  if (digitsToShow) {
+    const beginDigits: number = Math.ceil(digitsToShow / 2)
+    const endDigits: number = Math.floor(digitsToShow / 2)
+    const begin: string = address.substr(0, 2 + beginDigits)
+    const end: string = address.substr(address.length - endDigits, endDigits)
+    return `${begin}…${end}`
+  }
+  return address
+}
